enabled JavaScalaActors for sbt 0.11.2 and provided the neccessary 2.9.1 scala libs for a pure java eclipse project
