+++ chap08/JavaScalaActors/.classpath
@@ -1,8 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="src" path="src/main/resources"/>
+	<classpathentry kind="src" path="src/main/java"/>
+	<classpathentry kind="src" path="src/test/java"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.6"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/Scala"/>
-	<classpathentry kind="lib" path="/Users/n/mybook/code/chap08/scala_actors_wrapper/target/scala_2.8.1/scalaactorwrapper_2.8.1-1.0.jar"/>
-	<classpathentry kind="output" path="bin"/>
+	<classpathentry kind="lib" path="lib/scala-library_2.9.1.jar"/>
+	<classpathentry kind="lib" path="lib/scalaactorwrapper_2.9.1-1.0.jar"/>
+	<classpathentry kind="output" path="target/scala-2.9.1/classes"/>
 </classpath>
