--- chap08/JavaScalaActors/bin/main/resources/A.txt
@@ -1,2 +1,0 @@
-threading is hard
-threading is hard
