--- chap08/JavaScalaActors/bin/main/resources/B.txt
@@ -1,2 +1,0 @@
-I need higher abstraction layer
-I need higher abstraction layer
