--- chap08/JavaScalaActors/bin/main/resources/C.txt
@@ -1,2 +1,0 @@
-Actors are cool
-Actors are cool
