--- chap08/JavaScalaActors/bin/main/resources/D.txt
@@ -1,2 +1,0 @@
-Actors with futures are right way to share data
-Actors with futures are right way to share data
