+++ chap08/JavaScalaActors/build.sbt
@@ -1,0 +1,11 @@
+name := "JavaScalaActors"
+
+version := "1.0"
+
+organization := "Scala in Action"
+
+javacOptions ++= Seq("-source", "1.6")
+
+
+
+
