+++ chap08/JavaScalaActors/project/build.properties
@@ -1,0 +1,3 @@
+#Project properties
+#Mon Dec 13 23:56:52 EST 2010
+sbt.version=0.11.2
